+++ engine/src/render/ring_buffer.h
@@ -17,16 +17,16 @@ struct RingBuffer
 {
     std::string name;
     usize size = 0;
-    u32 offset = 0;
+    usize offset = 0;
     u32 usage = 0;
-    u32 last_frame_end = 0;
-    u32 last_frame_size = 0;
-    u32 this_frame_size = 0;
+    usize last_frame_end = 0;
+    usize last_frame_size = 0;
+    usize this_frame_size = 0;
     Handle<gfx::Buffer> buffer;
     bool should_align = true;
 
     static RingBuffer create(gfx::Device &device, const RingBufferDescription &desc, bool align = true);
-    std::pair<void*, u32> allocate(gfx::Device &device, usize len);
+    std::pair<void*, usize> allocate(gfx::Device &device, usize len);
     void start_frame();
     void end_frame();
 };
