+++ engine/src/render/unified_buffer_storage.cpp
@@ -3,7 +3,7 @@
 #include "render/vulkan/device.h"
 
 
-UnifiedBufferStorage UnifiedBufferStorage::create(gfx::Device &device, std::string name, u32 size_in_bytes, u32 element_size, u32 gpu_usage)
+UnifiedBufferStorage UnifiedBufferStorage::create(gfx::Device &device, std::string name, usize size_in_bytes, u32 element_size, u32 gpu_usage)
 {
     UnifiedBufferStorage storage = {};
     storage.allocator = BuddyAllocator::create(size_in_bytes);
@@ -16,7 +16,7 @@ UnifiedBufferStorage UnifiedBufferStorage::create(gfx::Device &device, std::stri
     return storage;
 }
 
-u32 UnifiedBufferStorage::allocate(u32 nb_element)
+u32 UnifiedBufferStorage::allocate(usize nb_element)
 {
     u32 byte_offset = allocator.allocate(nb_element * this->element_size);
     assert(byte_offset % this->element_size == 0);
