+++ engine/src/render/vulkan/descriptor_set.h
@@ -26,7 +26,7 @@ struct DynamicDescriptor
 {
     Handle<Buffer> buffer_handle;
     usize          size;
-    u32            offset;
+    usize          offset;
 };
 
 struct DescriptorType
@@ -78,13 +78,13 @@ struct DescriptorSet
 
     // dynamic offsets
     Vec<usize> dynamic_descriptors;
-    Vec<u32>   dynamic_offsets;
+    Vec<usize> dynamic_offsets;
 };
 
 DescriptorSet create_descriptor_set(Device &device, const Vec<DescriptorType> &descriptors);
 void          destroy_descriptor_set(Device &device, DescriptorSet &set);
 
-void bind_uniform_buffer(DescriptorSet &set, u32 slot, Handle<Buffer> buffer_handle, u32 offset, usize size);
+void bind_uniform_buffer(DescriptorSet &set, u32 slot, Handle<Buffer> buffer_handle, usize offset, usize size);
 void bind_storage_buffer(DescriptorSet &set, u32 slot, Handle<Buffer> buffer_handle);
 void bind_image(DescriptorSet &set, u32 slot, Handle<Image> image_handle);
